requirements.txt
Flask==3.1.1
numpy==1.24.3
Pillow==11.2.1
tensorflow==2.13.0
gunicorn==21.2.0
tensorflow_intel